moves components related to the formatting and display of variables into a variables folder
