+++ packages/cms/src/components/variables/ConsoleVariable.css
@@ -1,4 +1,4 @@
-@import "../css/mixins.css";
+@import "../../css/mixins.css";
 
 .cms-variable {
   color: inherit;
